remove 40_read_file.js and 40_write_file.js
