+++ runtime/build.rs
@@ -207,11 +207,9 @@ mod startup_snapshot {
         "40_fs_events.js",
         "40_http.js",
         "40_process.js",
-        "40_read_file.js",
         "40_signals.js",
         "40_spawn.js",
         "40_tty.js",
-        "40_write_file.js",
         "41_prompt.js",
         "90_deno_ns.js",
         "98_global_scope.js",
